Billing module and error issue fix -20-03-21

+++ Swayam/src/main/java/sbi/kiosk/swayam/common/dto/BillingPenaltyDto.java
@@ -67,6 +67,8 @@ public class BillingPenaltyDto {
 	private Integer relaxation;
 	private Integer finalDowntime;
 	
-	private Float penalty;
+//	private Float penalty;
+	
+	private Double penalty;
 	private String prfRefNumber;
 }
